remove StoredUserProfile interface
