+++ services/src/main/java/org/keycloak/authentication/authenticators/broker/IdpReviewProfileAuthenticator.java
@@ -37,9 +37,9 @@ import org.keycloak.services.resources.AttributeFormDataProcessor;
 import org.keycloak.services.validation.Validation;
 import org.keycloak.userprofile.LegacyUserProfileProviderFactory;
 import org.keycloak.userprofile.UserProfileProvider;
-import org.keycloak.userprofile.profile.represenations.AttributeUserProfile;
+import org.keycloak.userprofile.profile.representations.AttributeUserProfile;
 import org.keycloak.userprofile.profile.DefaultUserProfileContext;
-import org.keycloak.userprofile.profile.represenations.IdpUserProfile;
+import org.keycloak.userprofile.profile.representations.IdpUserProfile;
 import org.keycloak.userprofile.utils.UserProfileUpdateHelper;
 import org.keycloak.userprofile.validation.UserProfileValidationResult;
 import org.keycloak.userprofile.validation.UserUpdateEvent;
