+++ services/src/main/java/org/keycloak/services/resources/account/AccountRestService.java
@@ -22,7 +22,6 @@ import org.keycloak.common.ClientConnection;
 import org.keycloak.common.enums.AccountRestApiVersion;
 import org.keycloak.common.Profile;
 import org.keycloak.common.util.StringPropertyReplacer;
-import org.keycloak.credential.CredentialModel;
 import org.keycloak.events.EventBuilder;
 import org.keycloak.events.EventStoreProvider;
 import org.keycloak.events.EventType;
@@ -51,13 +50,11 @@ import org.keycloak.theme.Theme;
 import org.keycloak.userprofile.LegacyUserProfileProviderFactory;
 import org.keycloak.userprofile.UserProfileProvider;
 import org.keycloak.userprofile.utils.UserProfileUpdateHelper;
-import org.keycloak.userprofile.profile.represenations.AccountUserRepresentationUserProfile;
+import org.keycloak.userprofile.profile.representations.AccountUserRepresentationUserProfile;
 import org.keycloak.userprofile.profile.DefaultUserProfileContext;
-import org.keycloak.userprofile.profile.represenations.UserModelUserProfile;
+import org.keycloak.userprofile.profile.representations.UserModelUserProfile;
 import org.keycloak.userprofile.validation.UserProfileValidationResult;
 import org.keycloak.userprofile.validation.UserUpdateEvent;
-import org.keycloak.common.Profile;
-import org.keycloak.theme.Theme;
 
 import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
