+++ services/src/main/java/org/keycloak/userprofile/profile/representations/AttributeUserProfile.java
@@ -15,43 +15,37 @@
  * limitations under the License.
  */
 
-package org.keycloak.userprofile.profile.represenations;
+package org.keycloak.userprofile.profile.representations;
 
-import org.keycloak.models.UserModel;
 import org.keycloak.userprofile.profile.AbstractUserProfile;
 
+import javax.ws.rs.NotSupportedException;
 import java.util.List;
 import java.util.Map;
 
 /**
  * @author <a href="mailto:markus.till@bosch.io">Markus Till</a>
  */
-public class UserModelUserProfile extends AbstractUserProfile {
+public class AttributeUserProfile extends AbstractUserProfile {
+    private final Map<String, List<String>> attributes;
 
-    private final UserModel user;
 
-    public UserModelUserProfile(UserModel user) {
-        this.user = user;
-    }
-
-    @Override
-    public String getId() {
-        return user.getId();
+    public AttributeUserProfile(Map<String, List<String>> attributes) {
+        this.attributes = attributes;
     }
 
     @Override
     public Map<String, List<String>> getAttributes() {
-        return user.getAttributes();
+        return this.attributes;
     }
 
     @Override
     public void setAttribute(String key, List<String> value) {
-        user.setAttribute(key, value);
+        this.getAttributes().put(key, value);
     }
 
     @Override
-    public void removeAttribute(String attr) {
-        // Due to the fact that the user attribute list is a copy and not a reference in the user adapter we have to access the remove function directly
-        user.removeAttribute(attr);
+    public String getId() {
+        throw new NotSupportedException("No ID support");
     }
 }
